+++ src/OakLab.Blazor.Navigation/RouteTemplateToken.cs
@@ -1,0 +1,6 @@
+﻿namespace OakLab.Blazor.Navigation;
+
+internal readonly record struct RouteTemplateToken(string Value, int SeparatorCount)
+{
+    public override string ToString() => Value;
+}